requirements.txt
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.3
bottle==0.13.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
decorator==4.4.2
dill==0.4.1
filelock==3.20.3
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-speech==2.36.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
html2pic==0.2.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
movielite==0.2.2
moviepy==1.0.3
mpmath==1.3.0
multiprocess==0.70.19
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
opencv-python==4.12.0.88
pictex==2.0.1
pillow==12.1.0
playwright==1.57.0
proglog==0.1.12
proto-plus==1.27.0
protobuf==6.33.4
proxy_tools==0.1.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybind11==3.0.1
pycaps @ git+https://github.com/francozanardi/pycaps.git@d30db93102693f5f8b5c53ebd7be4333db8c61c6
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
python-dotenv==1.2.1
pywebview==6.1
regex==2026.1.15
requests==2.32.5
rich==14.2.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
skia-python==138.0
soupsieve==2.8.3
stretchable==1.1.8
sympy==1.14.0
tiktoken==0.12.0
tinycss2==1.5.1
torch==2.9.1
tqdm==4.67.1
triton==3.5.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
